+++ packages/discovery-react-components/src/components/DocumentPreview/__stories__/DocumentPreview.stories.tsx
@@ -3,6 +3,7 @@ import { storiesOf } from '@storybook/react';
 import { radios } from '@storybook/addon-knobs';
 import { QueryResult, QueryResultPassage } from 'ibm-watson/discovery/v2';
 import DocumentPreview from '../DocumentPreview';
+import { document as docPDF } from '../__fixtures__/Art Effects.pdf';
 import docArtEffects from '../__fixtures__/Art Effects Koya Creative Base TSA 2008.pdf.json';
 import passages from '../__fixtures__/passages';
 import jsonPassages from '../__fixtures__/jsonPassages';
@@ -17,36 +18,69 @@ const Wrapper: FC<WrapperProps> = ({ children, style = {} }) => (
   </div>
 );
 
-const doc = docArtEffects;
-const docWithoutHtml = omit(docArtEffects, 'html');
-
 storiesOf('DocumentPreview', module)
   .addParameters({ component: DocumentPreview })
   .add('default', () => {
+    const [file, doc] = docSelection();
     return (
       <Wrapper>
-        <DocumentPreview document={docWithoutHtml} />
-      </Wrapper>
-    );
-  })
-  .add('with Html', () => {
-    return (
-      <Wrapper>
-        <DocumentPreview document={doc} />
+        <DocumentPreview document={doc} file={file} />
       </Wrapper>
     );
   })
   .add('passage highlighting', () => {
+    const [file, doc] = docSelection();
     const usedPassage = doc.extracted_metadata.file_type === 'json' ? jsonPassages : passages;
     const docWithPassage = passageSelection(doc, usedPassage);
     const highlight = (docWithPassage.document_passages as unknown as QueryResultPassage[])[0];
 
     return (
       <Wrapper>
-        <DocumentPreview document={docWithPassage} highlight={highlight} />
+        <DocumentPreview document={docWithPassage} highlight={highlight} file={file} />
+      </Wrapper>
+    );
+  })
+  .add('table highlight', () => {
+    const [file, doc] = docSelection();
+    const docWithTable = tableSelection(doc);
+    const highlight = docWithTable.table_results[0];
+
+    return (
+      <Wrapper>
+        <DocumentPreview file={file} document={docWithTable} highlight={highlight} />
       </Wrapper>
     );
   });
+
+function docSelection(): [string | undefined, QueryResult] {
+  const label = 'Document Type';
+  const options = {
+    PDF: 'pdf',
+    'Document with `html` property and structure data': 'html',
+    'Document with only text': 'text'
+  };
+  const defaultValue = 'pdf';
+  const groupId = 'GROUP-ID1';
+  const docname = radios(label, options, defaultValue, groupId);
+
+  let file, doc;
+  switch (docname) {
+    case 'pdf':
+      file = atob(docPDF);
+      doc = docArtEffects;
+      break;
+    case 'html':
+      doc = docArtEffects;
+      break;
+    case 'text':
+      doc = omit(docArtEffects, 'html');
+      break;
+    default:
+      throw new Error('Unknown radios option');
+  }
+
+  return [file, doc];
+}
 
 function passageSelection(doc: QueryResult, passages: object): QueryResult {
   const label = 'Passage Type';
@@ -62,5 +96,28 @@ function passageSelection(doc: QueryResult, passages: object): QueryResult {
   return {
     ...doc,
     document_passages: [passages[passageType]]
+  };
+}
+
+function tableSelection(doc: QueryResult): QueryResult {
+  // add both table results and passages
+  return {
+    ...doc,
+    document_passages: [passages.single, passages.multiline],
+    table_results: [
+      {
+        table_id: '558ada041262d5b0aa02a05429d798c7',
+        source_document_id: '7e8ada041262d5b0aa02a05429d798c7',
+        collection_id: '8713a92b-28aa-b291-0000-016ddc68aa2a',
+        table_html: '<table><tr><th>Hello</th><tr><td>How are ya?</td></tr></table>',
+        table_html_offset: 42500,
+        table: {
+          location: {
+            begin: 346183,
+            end: 349624
+          }
+        }
+      }
+    ]
   };
 }